Set up different build targets
Set up a way to build for both Arduboy and M5. No m5 implementation
included yet, this is just getting the build setup in place to make it
easy to build either variant.

arduino-cli doesn't give any way to control which files will be included
in a build; it just includes everything in the root dir, and in src.

 So the Makefile sets up symbolic links to the
appropriate files before building for a particular target.